improvements and refactoring
Improvements done on the lending contracts:

changed borrowings variable to userBorrowData and lastBorrowTime field name to userLastTimestamp in the Borrow struct.
Added a updateAccountAndBorrowState that is called at the start of borrow and repay functions.
changed variables manipulation with proper handling of the token decimals being manipulated.
moved transfers at the end of all the functions.
removed the debtAmount field from the Account structure and now use only principal. and accruedInterest from the Borrow structure.
changed the calculateMaxBorrow function to return the value that is set in the calling function without updating the state inside.
liquidation function has been added.
added GRACE_THRESHOLD mechanism to set the debt and accrued interest to 0 when the (repaid amount - debt) is a few wei. This is to cope with cases of very small accrued interests accumulated between the approve and the repay calls.

Improvements done on the lending front end:

- Break down App.js into modular components
- Display values fixed
- added .env for configuration
- bug fixes